+++ CHANGES.rst
@@ -1,8 +1,11 @@
 Changelog
 ---------
 
-0.9.1 (unreleased)
-~~~~~~~~~~~~~~~~~~
+0.10 (unreleased)
+~~~~~~~~~~~~~~~~~
+
+- ``%filter`` is disabled by default for being a security risk.  Use
+  ``--unsafe`` to enable.
 
 
 0.9 (2014-05-09)